add ProfilesLogicTest
